[LM-83] Split server.py into server/ package (routes, models, db, helpers)
