+++ run.sh
@@ -7,4 +7,4 @@ set -euo pipefail
 SCRIPT_DIR="$(cd "$(dirname "$0")" && pwd)"
 cd "$SCRIPT_DIR"
 
-exec python3 -m uvicorn server:app --host 127.0.0.1 --port 18792
+exec python3 -m uvicorn server.app:app --host 127.0.0.1 --port 18792
