+++ server/__init__.py
@@ -1,0 +1,6 @@
+from server import db  # noqa: F401
+from server.db import apply_pending_migrations, get_db  # noqa: F401
+from server.models import ReportPayload, VerdictPayload  # noqa: F401
+from server.constants import MONITOR_VERSION, PROJECTS, SUPPORTED_API_MAJOR  # noqa: F401
+from server.routes.ingest import _insert_event, _insert_verdict  # noqa: F401
+from server.app import app  # noqa: F401
