+++ server/routes/health.py
@@ -1,0 +1,48 @@
+from fastapi import APIRouter
+from server.db import get_db
+from server.constants import MONITOR_VERSION, SUPPORTED_API_MAJOR
+
+router = APIRouter()
+
+
+@router.get("/api/health")
+def health():
+    conn = get_db()
+    rows = conn.execute(
+        "SELECT core_version, COUNT(*) as cnt FROM events WHERE core_version IS NOT NULL GROUP BY core_version"
+    ).fetchall()
+    core_version_counts = {r["core_version"]: r["cnt"] for r in rows}
+    loop_rows = conn.execute(
+        "SELECT DISTINCT COALESCE(loop_id, '(unknown)') AS loop_id FROM events ORDER BY loop_id"
+    ).fetchall()
+    loop_ids = [r["loop_id"] for r in loop_rows]
+    conn.close()
+    return {
+        "status": "ok",
+        "monitor_version": MONITOR_VERSION,
+        "supported_bounty_api": f"{SUPPORTED_API_MAJOR}.x",
+        "core_version_counts": core_version_counts,
+        "loop_ids": loop_ids,
+    }
+
+
+@router.get("/api/loops")
+def get_loops():
+    conn = get_db()
+    rows = conn.execute("""
+        SELECT COALESCE(loop_id, '(unknown)') AS loop_id,
+               MAX(created_at) AS last_seen,
+               COUNT(*) AS event_count,
+               GROUP_CONCAT(DISTINCT core_version) AS core_versions
+        FROM events
+        GROUP BY COALESCE(loop_id, '(unknown)')
+        ORDER BY 1
+    """).fetchall()
+    conn.close()
+    result = []
+    for r in rows:
+        entry = dict(r)
+        raw = entry.get("core_versions") or ""
+        entry["core_versions"] = sorted(set(v for v in raw.split(",") if v)) if raw else []
+        result.append(entry)
+    return result
